+++ scratch/pilot_sample.py
